First implementation ,will split the logic by feature
