+++ BuberDinner/BuberDinner.Api/Controllers/AuthenticationController.cs
@@ -1,23 +1,28 @@
-using BuberDinner.Application.Services.Authentication;
 using BuberDinner.Contracts.Authentication;
 using ErrorOr;
 using Microsoft.AspNetCore.Mvc;
-using BuberDinner.Domain.Common.Errors;
+using BuberDinner.Application.Services.Authentication.Commands;
+using BuberDinner.Application.Services.Authentication.Queries;
+using BuberDinner.Application.Services.Authentication.Common;
 namespace BuberDinner.Api.Controllers;
 
 
 [Route("auth")]
 public class AuthenticationController: ApiController 
 {
-    private readonly IAuthenticationService _authenticationService;
+    private readonly IAuthenticationCommandService _authenticationCommandService;
+    private readonly IAuthenticationQueryService _authenticationQueryService;
 
-    public AuthenticationController(IAuthenticationService authenticationService)
+
+    public AuthenticationController(IAuthenticationCommandService authenticationCommandService,
+                                    IAuthenticationQueryService authenticationQueryService)
     {
-        _authenticationService = authenticationService;
+        _authenticationCommandService = authenticationCommandService;
+        _authenticationQueryService = authenticationQueryService;
     }
     [HttpPost("register")]
     public IActionResult Register(RegisterRequest request){
-        ErrorOr<AuthenticationResult> authResult = _authenticationService.Register(
+        ErrorOr<AuthenticationResult> authResult = _authenticationCommandService.Register(
             request.FirstName,
             request.LastName,
             request.Email,
@@ -42,7 +47,7 @@ public class AuthenticationController: ApiController
 
     [HttpPost("login")]
     public IActionResult Login(LoginRequest request){
-         var authResult = _authenticationService.Login(
+         var authResult = _authenticationQueryService.Login(
             request.Email,
             request.Password
         );
