+++ BuberDinner/BuberDinner.Application/DependencyInjection.cs
@@ -1,12 +1,13 @@
-using BuberDinner.Application.Services.Authentication;
+using BuberDinner.Application.Services.Authentication.Commands;
+using BuberDinner.Application.Services.Authentication.Queries;
 using Microsoft.Extensions.DependencyInjection;
-
 namespace BuberDinner.Application;
 
 public static class DependencyInjection {
     public static IServiceCollection AddApplication( this IServiceCollection services)
-    {
-            services.AddScoped<IAuthenticationService, AuthenticationService>();
+    {   
+            services.AddScoped<IAuthenticationCommandService,AuthenticationCommandService>();
+            services.AddScoped<IAuthenticationQueryService, AuthenticationQueryService>();
             return services;
     }
 }
