+++ BuberDinner/BuberDinner.Application/Services/Authentication/Commands/AuthenticationCommandService.cs
@@ -1,42 +1,23 @@
-using BuberDinner.Application.Common.Errors;
 using BuberDinner.Application.Common.Interfaces.Authentication;
 using BuberDinner.Application.Common.Interfaces.Persistance;
+using BuberDinner.Application.Services.Authentication.Common;
 using BuberDinner.Domain.Common.Errors;
 using BuberDinner.Domain.Entities;
 using ErrorOr;
 
-namespace BuberDinner.Application.Services.Authentication;
+namespace BuberDinner.Application.Services.Authentication.Commands;
 
-public class AuthenticationService : IAuthenticationService
+public class AuthenticationCommandService : IAuthenticationCommandService
 {
     private readonly IJwtTokenGenerator _jwtTokenGenerator;
     private readonly IUserRepository _userRepository;
 
-    public AuthenticationService(IJwtTokenGenerator jwtTokenGenerator,IUserRepository userRepository)
+    public AuthenticationCommandService(IJwtTokenGenerator jwtTokenGenerator,IUserRepository userRepository)
     {
         _jwtTokenGenerator = jwtTokenGenerator;
         _userRepository = userRepository;
         
     }
-    public ErrorOr<AuthenticationResult> Login(string email, string password)
-    {
-        // 1. Validate the user exists
-        if(_userRepository.GetUserByEmail(email) is not User user)
-        {
-            return Errors.Authentication.InvalidCredentials;
-        }
-        // 2. Validate the password is correct
-        if(user.Password!=password)
-        {
-            return new [] {Errors.Authentication.InvalidCredentials};
-        }
-        // 3. Create JWT token
-        var token = _jwtTokenGenerator.GenerateToken(user);
-        return new AuthenticationResult(
-         user,
-            token);
-    }
-
     public ErrorOr<AuthenticationResult> Register(string firstName, string lastName, string email, string password)
     {
         // 1. Validate the user doesn't exist
