+++ BuberDinner/BuberDinner.Application/Services/Authentication/Commands/IAuthenticationCommandService.cs
@@ -1,0 +1,8 @@
+using BuberDinner.Application.Services.Authentication.Common;
+using ErrorOr;
+
+namespace BuberDinner.Application.Services.Authentication.Commands;
+public interface IAuthenticationCommandService 
+{
+    ErrorOr<AuthenticationResult> Register(string firstName,string lastName,string email, string password);
+}
