+++ BuberDinner/BuberDinner.Application/Services/Authentication/Common/AuthenticationResult.cs
@@ -1,6 +1,6 @@
 using BuberDinner.Domain.Entities;
 
-namespace  BuberDinner.Application.Services.Authentication;
+namespace  BuberDinner.Application.Services.Authentication.Common;
 
 public record AuthenticationResult (
     User User,
