--- BuberDinner/BuberDinner.Application/Services/Authentication/IAuthenticationsService.cs
@@ -1,10 +1,0 @@
-
-
-using ErrorOr;
-
-namespace BuberDinner.Application.Services.Authentication;
-public interface IAuthenticationService 
-{
-    ErrorOr<AuthenticationResult> Login(string email,string password);
-    ErrorOr<AuthenticationResult> Register(string firstName,string lastName,string email, string password);
-}
