+++ BuberDinner/BuberDinner.Application/Services/Authentication/Queries/IAuthenticationQueryService.cs
@@ -1,0 +1,8 @@
+using BuberDinner.Application.Services.Authentication.Common;
+using ErrorOr;
+
+namespace BuberDinner.Application.Services.Authentication.Queries;
+public interface IAuthenticationQueryService 
+{
+    ErrorOr<AuthenticationResult> Login(string email,string password);
+}
